🏗️ sturcture: structure change
